+++ tests/tests/wgpu-gpu/zero_init.rs
@@ -455,7 +455,9 @@ static WRITE_TEXTURE_PLANE0_LEAVES_PLANE1_UNINIT_NV12: GpuTestConfiguration =
         .parameters(
             TestParameters::default()
                 .features(Features::TEXTURE_FORMAT_NV12)
-                .limits(Limits::downlevel_defaults()),
+                .limits(Limits::downlevel_defaults())
+                // https://github.com/gfx-rs/wgpu/issues/9688
+                .expect_fail(FailureCase::lvp_poison_memory("read back non-zero")),
         )
         .run_async(|ctx| async move {
             check_plane_write_leaves_other_plane_uninit(
@@ -475,7 +477,9 @@ static WRITE_TEXTURE_PLANE1_LEAVES_PLANE0_UNINIT_NV12: GpuTestConfiguration =
         .parameters(
             TestParameters::default()
                 .features(Features::TEXTURE_FORMAT_NV12)
-                .limits(Limits::downlevel_defaults()),
+                .limits(Limits::downlevel_defaults())
+                // https://github.com/gfx-rs/wgpu/issues/9688
+                .expect_fail(FailureCase::lvp_poison_memory("read back non-zero")),
         )
         .run_async(|ctx| async move {
             check_plane_write_leaves_other_plane_uninit(
@@ -495,7 +499,9 @@ static WRITE_TEXTURE_PLANE0_LEAVES_PLANE1_UNINIT_P010: GpuTestConfiguration =
         .parameters(
             TestParameters::default()
                 .features(Features::TEXTURE_FORMAT_P010 | Features::TEXTURE_FORMAT_16BIT_NORM)
-                .limits(Limits::downlevel_defaults()),
+                .limits(Limits::downlevel_defaults())
+                // https://github.com/gfx-rs/wgpu/issues/9688
+                .expect_fail(FailureCase::lvp_poison_memory("read back non-zero")),
         )
         .run_async(|ctx| async move {
             check_plane_write_leaves_other_plane_uninit(
@@ -515,7 +521,9 @@ static WRITE_TEXTURE_PLANE1_LEAVES_PLANE0_UNINIT_P010: GpuTestConfiguration =
         .parameters(
             TestParameters::default()
                 .features(Features::TEXTURE_FORMAT_P010 | Features::TEXTURE_FORMAT_16BIT_NORM)
-                .limits(Limits::downlevel_defaults()),
+                .limits(Limits::downlevel_defaults())
+                // https://github.com/gfx-rs/wgpu/issues/9688
+                .expect_fail(FailureCase::lvp_poison_memory("read back non-zero")),
         )
         .run_async(|ctx| async move {
             check_plane_write_leaves_other_plane_uninit(
@@ -558,7 +566,9 @@ static COPY_BUFFER_TO_TEXTURE_PLANE0_LEAVES_PLANE1_UNINIT_NV12: GpuTestConfigura
         .parameters(
             TestParameters::default()
                 .features(Features::TEXTURE_FORMAT_NV12)
-                .limits(Limits::downlevel_defaults()),
+                .limits(Limits::downlevel_defaults())
+                // https://github.com/gfx-rs/wgpu/issues/9688
+                .expect_fail(FailureCase::lvp_poison_memory("read back non-zero")),
         )
         .run_async(|ctx| async move {
             check_plane_write_leaves_other_plane_uninit(